backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
pydantic-settings
redis
celery[redis]
graphviz